LayoutTests/media/video-pause-empty-events-expected.txt

Test that pause() from EMPTY network state triggers load()

TEST(video.networkState == HTMLMediaElement.EMPTY) OK
RUN(video.pause())
EVENT(loadstart)
END OF TEST
